refactor(pages): move 10 oversized single-file pages into kebab-case folders
Per CLAUDE.md convention: multi-file pages (>300 lines) live in
kebab-case directories with their own components/ subfolder.

Moved pages (all above the 300-line threshold):
  HomePage.tsx             -> home/
  BudgetPage.tsx           -> budget/
  NetWorthPage.tsx         -> net-worth/
  InvestmentAnalyticsPage  -> investment-analytics/
  MutualFundProjection     -> mutual-fund-projection/
  ReturnsAnalysisPage      -> returns-analysis/
  SpendingAnalysisPage     -> spending-analysis/
  UploadSyncPage           -> upload-sync/
  IncomeAnalysisPage       -> income-analysis/
  GSTAnalysisPage          -> gst-analysis/

App.tsx lazy-import paths updated to point at the new locations.
Each folder pre-seeded with an empty components/ subfolder for future
sub-component extraction. No behavioural changes.
