+++ frontend/src/pages/income-analysis/IncomeAnalysisPage.tsx
