+++ frontend/src/pages/investment-analytics/InvestmentAnalyticsPage.tsx
