+++ frontend/src/pages/mutual-fund-projection/MutualFundProjectionPage.tsx
