+++ frontend/src/pages/net-worth/NetWorthPage.tsx
