+++ frontend/src/pages/spending-analysis/SpendingAnalysisPage.tsx
