+++ frontend/src/pages/upload-sync/UploadSyncPage.tsx
